refactor: separate display and indexing layers with proper architecture
Split load.js (602 lines) into three focused modules aligned with
mldecouple architecture:

**New Structure:**
- display/data.js (295 lines) - Load & process index for UI display
- indexing/build.js (280 lines) - Orchestrate index builds via workers
- indexing/locks.js (100 lines) - Manage index build locks

**Renamed:**
- core/indexing-bridge.js → core/indexing-adapter.js (clearer name)

**Benefits:**
- Clear separation: display layer reads index, indexing layer builds it
- No circular dependencies
- Easier testing and maintenance
- Aligns with event-based architecture (adapter translates events → UI)

**Architecture:**
```
display/data.js → reads published index for UI
core/indexing-adapter.js → translates indexing events → UI state
indexing/build.js → builds index, emits events
indexing/locks.js → manages build locks
```

All 580 tests passing ✓
