+++ nx/blocks/media-library/core/indexing-adapter.js
